generators/python/pyproject.toml
[tool.poetry]
name = "fern_python"
version = "0.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.8.2"
typer = {extras = ["all"], version = "^0.6.1"}
fern-fern-generator-exec-sdk = {version = "0.82.5", source = "fern-prod"}
ordered-set = "^4.1.0"
pydantic-core = "^2.18.2"
fern-fern-fdr-sdk = {version = "0.98.20", source = "fern-prod"}
fern-fern-generator-cli-sdk = {version = "0.0.74", source = "fern-prod"}
fern_fern_ir_v53 = "53.12.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
mypy = "^0.971"
black = "^24.3.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
pre-commit = "^2.20.0"
snapshottest = "^0.6.0"

[[tool.poetry.source]]
name = "fern-prod"
url = "https://pypi.buildwithfern.com/"
default = false
secondary = true

[tool.poetry.group.dev.dependencies]
pytest-only = "^2.0.0"
ruff = "^0.5.6"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'


[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = [ "tests" ]
addopts = "--ignore-glob=tests/**/generated --ignore-glob=tests/**/snapshots"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
